% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandLetter.R
\name{RandLetter}
\alias{RandLetter}
\title{Randomly Choose a Number of Letters}
\usage{
RandLetter(n=1)
}
\arguments{
\item{n}{number of letters}
}
\value{
A number of letters
}
\description{
Randomly Choose a Number of Letters
}
\author{
Prayogi, Egi
}
